Change matrix multiplication into "evolve" method
Was having issues with matrix multiplication supplying the wrong type to
the product. (y = Ax gives y the type of A, not x.) So the evolve method
makes the types correct and also gives A the ability to specify the
destination type. E.g. oscillation probability turns a beam flux into a
beam flux, but cross section turns the beam flux into a spectrum.