Improve extract and zipWithEnergy + 1 hack (below)
For some reason, the SimulationComponent __new__ method is returning
objects of type numpy.matrix instead of type "cls" when the data
supplied is in block matrix filename form. To remedy this, I
have to "view cast" the data that is fed into the numpy.matrix __new__
method. I don't think I should have to do this, but it shouldn't have
any ill effects. (famous last words, but that's life)