Improve handling of filename/block matrices
Each class has a "_getBlockMatrixForm" method which takes in an
array/list of filenames and outputs those filenames in the correct block
matrix format. This allows e.g. different behavior for a column vector
input for cross sections (ends up as a matrix) vs. flux (stays as a
column vector).