Add option for factored/unfactored default DRM
"factored" first does an energy reconstruction within each interaction
type, with the expectation that interaction channel ID will happen
afterwards and is mostly independent of energy reconstruction. It
answers the question "what are the chances a neutrino with energy Etrue,
interacting via <channel> is reconstructed as having energy Ereco,
assuming correctly that the interaction was a <channel> interaction?"

"unfactored" does both simultaneously, and answers the question "what
are the chances a neutrino with energy Etrue, channel <ch true> is
reconstructed as having energy Ereco and channel <ch reco>, with no
assumptions about the flavor or interaction channel.