Convert setup to a script that will be sourced
instead of executed. This mostly involved removing the `exit 0` line,
which would exit from the shell instead of from the script. I also
updated the README to remove the advice to `. setup` because the correct
command is `. ./setup` and it's just simpler to have one command that
works: `source setup`.