Make use of the repository location from "setup"
Instead of assuming a path to the repository, fetch the path from the
appropriate environment variable. That variable should exist as long as
the user sources the setup script first. The repository location is only
fetched if it is needed. If the repository location is not found, the
entire function exits with `sys.exit()` to force the user to figure out
what is going on.